+++ urtcsdk/URTCEngineImpl.cc
@@ -25,7 +25,11 @@ int URTCEngineImpl::InitRTCEngine(void* callback)
 	m_rtcengine->configLocalAudioPublish(URTCConfig::getInstance()->isAutoPubAudio());
 	m_rtcengine->configLocalCameraPublish(URTCConfig::getInstance()->isAutoPubVideo());
 	m_rtcengine->configLocalScreenPublish(URTCConfig::getInstance()->isAutoPubScreen());
-	m_rtcengine->setVideoProfile(URTCConfig::getInstance()->getVideoProfile());
+	tUCloudVideoConfig videoconfig;
+	videoconfig.mWidth = 0;
+	videoconfig.mHeight = 0;
+	videoconfig.mFrameRate = 0;
+	m_rtcengine->setVideoProfile(URTCConfig::getInstance()->getVideoProfile(), videoconfig);
 	m_rtcengine->setDesktopProfile(UCLOUD_RTC_SCREEN_PROFILE_HIGH_PLUS);
 	m_rtcengine->setSdkMode(URTCConfig::getInstance()->getSdkMode());
 	return 0;
@@ -297,6 +301,11 @@ int URTCEngineImpl::StartRecord(tRecordConfig& config)
 		uconfig.mWatermarkPos = (eUCloudRtcWaterMarkPos)config.mWatermarkPos;
 		uconfig.mBucket = "urtc-test";
 		uconfig.mBucketRegion = "cn-bj";
+
+		uconfig.mWaterMarkType = UCLOUD_RTC_WATERMARK_TYPE_TIME;
+		uconfig.mWatermarkUrl = "";
+		uconfig.mIsaverage = false;
+		uconfig.mMixerTemplateType = 9;
 		return m_rtcengine->startRecord(uconfig);
 	}
 	return -1;