+++ doc/changelog.md
@@ -258,11 +258,46 @@ m_rtcengine->setCaptureScreenPagrams(pgram);
 * 1 加入大班课功能，支持超大规模会议功能
 使用如下
 ``` c++ 
+必须在最开始调用
 m_rtcengine->setChannelType(UCLOUD_RTC_CHANNEL_TYPE_BROADCAST);
 注意：
 大班课中用户只能拥有订阅或者发布一种权限
 如果m_rtcengine->setStreamRole(UCLOUD_RTC_USER_STREAM_ROLE_BOTH);
 SDK会默认把权限设置为UCLOUD_RTC_USER_STREAM_ROLE_SUB
 ```
-* 2 添加超大分辨率支持1080p_plus
+
+# 1.6.1 版本发布
+* 1 加入大班课功能，支持超大规模会议功能
+使用如下
+``` c++ 
+必须在最开始调用
+m_rtcengine->setChannelType(UCLOUD_RTC_CHANNEL_TYPE_BROADCAST);
+注意：
+大班课中用户只能拥有订阅或者发布一种权限
+如果m_rtcengine->setStreamRole(UCLOUD_RTC_USER_STREAM_ROLE_BOTH);
+SDK会默认把权限设置为UCLOUD_RTC_USER_STREAM_ROLE_SUB
+```
+* 2 添加自定义编码（最大分辨率到1080p(1920*1080)）
+``` c++ 
+virtual void setVideoProfile(eUCloudRtcVideoProfile profile) = 0;
+变更为
+virtual void setVideoProfile(eUCloudRtcVideoProfile profile, tUCloudVideoConfig& videoconfig) = 0;
+tUCloudVideoConfig 用来定义自己编码分辨率
+```
+* 3 录制功能增加更多录制模板以及水印 调用示例如下
+``` c++ 
+tUCloudRtcRecordConfig recordconfig;
+recordconfig.mMainviewmediatype = UCLOUD_RTC_MEDIATYPE_VIDEO; // 主画面类型 video screen
+recordconfig.mMainviewuid = m_userid.data(); // 主画面用户id
+recordconfig.mProfile = UCLOUD_RTC_RECORDPROFILE_SD;//录制输出等级
+recordconfig.mRecordType = UCLOUD_RTC_RECORDTYPE_AUDIOVIDEO;//录制媒体内容
+recordconfig.mWatermarkPos = UCLOUD_RTC_WATERMARKPOS_LEFTTOP;//水印位置
+recordconfig.mBucket = "urtc-test";  // ufile bucket
+recordconfig.mBucketRegion = "cn-bj"; // ufile region
+recordconfig.mIsaverage = false; // 画面是否均分 不均分 均采用 1大几小格式 大画面在左 小画面在右
+recordconfig.mWaterMarkType = UCLOUD_RTC_WATERMARK_TYPE_TIME;  // 水印类型
+recordconfig.mWatermarkUrl = "hello urtc"; // 如果是文字水印为水印内容   如果是图片则为图片url 地址
+recordconfig.mMixerTemplateType = 4; [混流模板](http:https://github.com/UCloudDocs/urtc/blob/master/cloudRecord/RecordLaylout.md)
+m_rtcengine->startRecord(recordconfig);
+```